Clock control (#144)
Adding simulated clock management :
- Support for multiple internal & external clock sources
- the frequency argument in CLI becomes optional
- Adding Clock Controller models for AVR and XT architectures

+++ lib_core/src/core/sim_firmware.h
@@ -76,8 +76,8 @@ public:
         Area        area;
     };
 
-    ///Main clock frequency in hertz, mandatory to run the simulation.
-    unsigned long frequency;
+    ///Fixed clock frequency in hertz
+    double frequency;
     ///Power supply voltage in volts. If not set, analog peripherals such as ADC are not usable.
     double vcc;
     ///Analog reference voltage in volts
